move helpping fn to utils
