+++ lib/utils/outside-board.ts
